requirements.txt
streamlit==1.31.0
Pillow==10.2.0
requests==2.31.0
pyautogen==0.2.14
openai==1.12.0
httpx==0.26.0